@@ -1,16 +1,49 @@
 module github.com/Safecast/TTServe
 
-go 1.12
+go 1.21
+
+toolchain go1.23.0
 
 require (
 	github.com/Safecast/safecast-go v0.0.0-20210602192847-dcf96c859f9b
-	github.com/Safecast/ttproto v0.0.0-20180220033412-d66a002deafc
+	github.com/Safecast/ttproto v0.0.0-20241212144031-fb7784e66a8b
 	github.com/blues/note-go v1.7.2
-	github.com/eclipse/paho.mqtt.golang v1.3.5
-	github.com/golang/protobuf v1.5.2
+	github.com/eclipse/paho.mqtt.golang v1.5.0
+	github.com/golang/protobuf v1.5.4
+	github.com/google/open-location-code/go v0.0.0-20241206115931-b6937486c621
+	google.golang.org/protobuf v1.35.2
+)
+
+require (
+	github.com/StackExchange/wmi v0.0.0-20190523213315-cbe66965904d // indirect
+	github.com/creack/goselect v0.1.2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-ole/go-ole v1.2.4 // indirect
+	github.com/gofrs/flock v0.7.1 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
-	github.com/google/open-location-code/go v0.0.0-20220120191843-cafb35c0d74d
-	github.com/gorilla/websocket v1.5.0 // indirect
-	golang.org/x/net v0.25.0 // indirect
-	google.golang.org/protobuf v1.28.0 // indirect
+	github.com/gorilla/websocket v1.5.3 // indirect
+	github.com/jonboulle/clockwork v0.3.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/shirou/gopsutil/v3 v3.21.6 // indirect
+	github.com/stretchr/objx v0.1.0 // indirect
+	github.com/stretchr/testify v1.7.0 // indirect
+	github.com/tklauser/go-sysconf v0.3.6 // indirect
+	github.com/tklauser/numcpus v0.2.2 // indirect
+	github.com/yuin/goldmark v1.4.13 // indirect
+	go.bug.st/serial v1.6.1 // indirect
+	golang.org/x/crypto v0.30.0 // indirect
+	golang.org/x/mod v0.17.0 // indirect
+	golang.org/x/net v0.32.0 // indirect
+	golang.org/x/sync v0.10.0 // indirect
+	golang.org/x/sys v0.28.0 // indirect
+	golang.org/x/telemetry v0.0.0-20240228155512-f48c80bd79b2 // indirect
+	golang.org/x/term v0.27.0 // indirect
+	golang.org/x/text v0.21.0 // indirect
+	golang.org/x/tools v0.21.1-0.20240508182429-e35e4ccd0d2d // indirect
+	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543 // indirect
+	gopkg.in/check.v1 v0.0.0-20161208181325-20d25e280405 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	periph.io/x/conn/v3 v3.7.0 // indirect
+	periph.io/x/d2xx v0.1.0 // indirect
+	periph.io/x/host/v3 v3.8.0 // indirect
 )
