@@ -9,7 +9,7 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"os"
 	"strings"
@@ -52,7 +52,7 @@ func ReadDeviceStatus(deviceUID string) (isAvail bool, isReset bool, sv DeviceSt
 	for i := 0; i < 5; i++ {
 
 		// Read the file and unmarshall if no error
-		contents, errRead := ioutil.ReadFile(filename)
+		contents, errRead := os.ReadFile(filename)
 		if errRead == nil {
 			valueToRead := DeviceStatus{}
 			errRead = json.Unmarshal(contents, &valueToRead)
@@ -838,7 +838,7 @@ func WriteDeviceStatus(sc ttdata.SafecastData) {
 			response, err := http.Get("http://ip-api.com/json/" + IP)
 			if err == nil {
 				defer response.Body.Close()
-				contents, err := ioutil.ReadAll(response.Body)
+				contents, err := io.ReadAll(response.Body)
 				if err == nil {
 					var info IPInfoData
 					err = json.Unmarshal(contents, &info)
