@@ -8,7 +8,7 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"os"
 )
@@ -19,7 +19,7 @@ func inboundWebGithubHandler(rw http.ResponseWriter, req *http.Request) {
 	stats.Count.HTTPGithub++
 
 	// Unpack the request
-	body, err := ioutil.ReadAll(req.Body)
+	body, err := io.ReadAll(req.Body)
 	if err != nil {
 		fmt.Printf("Github webhook: error reading body: %s\n", err)
 		return
